+++ app/Providers/AuthServiceProvider.php
@@ -6,9 +6,11 @@ use App\Models\Contact;
 use App\Models\KbArticle;
 use App\Models\Message;
 use App\Models\Ticket;
+use App\Models\TicketEvent;
 use App\Policies\ContactPolicy;
 use App\Policies\KbArticlePolicy;
 use App\Policies\MessagePolicy;
+use App\Policies\TicketEventPolicy;
 use App\Policies\TicketPolicy;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Gate;
@@ -17,6 +19,7 @@ class AuthServiceProvider extends ServiceProvider
 {
     protected $policies = [
         Ticket::class => TicketPolicy::class,
+        TicketEvent::class => TicketEventPolicy::class,
         Contact::class => ContactPolicy::class,
         KbArticle::class => KbArticlePolicy::class,
         Message::class => MessagePolicy::class,